Set website structure
Add directorys like recipes as webpages and images
for the whole site. My goal was to commit all directorys
and files together so i added one line to the README file,
which existed before. If i didnt do this, i couldnt
add the README file to the Stage and Commit.
But i don't know if this is the common way.